Create JWT util class

+++ src/main/java/com/nowakartur97/personalkanbanboardbackend/user/UserResponse.java
@@ -4,6 +4,5 @@ public record UserResponse(String username,
                            String password,
                            String email,
                            String token,
-                           int expirationTimeInMilliseconds
-) {
+                           long expirationTimeInMilliseconds) {
 }
